Added checking ig git_current_branch exist in git
Now we are checking if function/binary git_current_branch exist
in ggpf function from autoload_scripts/git.zsh

+++ autoload_scripts/git.zsh
@@ -73,6 +73,11 @@ ggpf() {
         return 1
     fi
 
+    if ! command -v git_current_branch >/dev/null 2>&1; then
+        echo "ggpf(): couldn't find function: git_current_branch" >&2
+        return 1
+    fi
+
     local branch; branch="$(git_current_branch)" || return
 
     git push --force origin "${branch}"
